added 404-page router replaced with switch-router and proper paths Folder structure updated
