+++ .gitignore
@@ -24,3 +24,6 @@ yarn-error.log*
 
 ../../.idea
 /.idea
+/.firebaserc
+/firebase.json
+/.firebase/hosting.YnVpbGQ.cache
