+++ src/Utils/actions/firestore.js
@@ -1,4 +1,4 @@
-import { ActionTypes } from '../config/constants'
+import { ActionTypes } from '../../config/constants'
 
 export const startDiaryNotesFirestoreSync = () => ({
   type: ActionTypes.START_FIRESTORE_DIARY_NOTES_SYNC,
