+++ src/Utils/reducers/login.js
@@ -1,4 +1,4 @@
-import { ActionTypes } from '../config/constants'
+import { ActionTypes } from '../../config/constants'
 
 const loginReducer = (state = false, action) => {
   switch (action.type) {
