+++ src/Utils/store.js
@@ -1,7 +1,7 @@
 import { applyMiddleware, createStore } from 'redux'
 import createSagaMiddleware from 'redux-saga'
-import rootReducer from '../reducers'
-import { watchSaga } from '../sagas/rootSaga'
+import rootReducer from './reducers'
+import { watchSaga } from './sagas/rootSaga'
 import { composeWithDevTools } from 'redux-devtools-extension'
 import { createFirestoreInstance } from 'redux-firestore'
 import { firebase, rrfConfig } from './firebase'
